fix(arc-742): centralise dates further and adjust notation on object detail
