--- src/modules/cp/utils/index.ts
@@ -1,1 +1,0 @@
-export * from './requests.utils';
