+++ src/styles/components/_button.scss
@@ -28,10 +28,6 @@
 		border-color: $teal-dark;
 	}
 
-	&:not(.c-button--no-height) {
-		height: 4.8rem;
-	}
-
 	/**
 		* Cross-browser styles for explicit focus via
 		* keyboard-based (eg Tab) navigation or the
@@ -48,6 +44,13 @@
 	&:focus:not(:focus-visible) {
 		outline: 0;
 		box-shadow: none;
+	}
+
+	/* stylelint-disable-next-line no-descending-specificity */
+	&:not(.c-button--no-height) {
+		&:not(.c-button--sm):not(.c-button--xs):not(.c-button--xxs) {
+			height: 4.8rem;
+		}
 	}
 }
